# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-Cloud Terrarium (mc-terrarium) aims to provide an environment to enrich multi-cloud infrastructure.",
        "title": "Multi-Cloud Terrarium REST API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/cloud-barista/mc-terrarium/issues/new"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.1.2"
    },
    "host": "localhost:8055",
    "basePath": "/terrarium",
    "paths": {
...
        "/tr": {
            "get": {
                "description": "Read all terrarium",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Terrarium] An environment to enrich the multi-cloud infrastructure"
                ],
                "summary": "Read all terrarium",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TerrariumInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Issue/create a terrarium",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Terrarium] An environment to enrich the multi-cloud infrastructure"
                ],
                "summary": "Issue/create a terrarium",
                "parameters": [
                    {
                        "description": "Information for a new terrarium",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TerrariumCreationRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}": {
            "get": {
                "description": "Read a terrarium",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Terrarium] An environment to enrich the multi-cloud infrastructure"
                ],
                "summary": "Read a terrarium",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TerrariumInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Erase the entire terrarium including directories and configuration files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Terrarium] An environment to enrich the multi-cloud infrastructure"
                ],
                "summary": "Erase the entire terrarium including directories and configuration files",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}/message-broker": {
            "get": {
                "description": "Get resource info of Message Broker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Message Broker] Operations (PoC - Not officially supported)"
                ],
                "summary": "Get resource info of Message Broker",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "refined",
                        "description": "Resource info by detail (refined, raw)",
                        "name": "detail",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Refresh the state before getting the info",
                        "name": "refresh",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Message Broker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Message Broker] Operations (PoC - Not officially supported)"
                ],
                "summary": "Create Message Broker",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Destroy Message Broker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Message Broker] Operations (PoC - Not officially supported)"
                ],
                "summary": "Destroy Message Broker",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/tr/{trId}/message-broker/infracode": {
            "post": {
                "description": "Create the infracode for Message Broker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Message Broker] Operations (PoC - Not officially supported)"
                ],
                "summary": "Create the infracode for Message Broker",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parameters of infracode for Message Broker",
                        "name": "ParamsForInfracode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateInfracodeOfMessageBrokerRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}/message-broker/plan": {
            "post": {
                "description": "Check and show changes by the current infracode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Message Broker] Operations (PoC - Not officially supported)"
                ],
                "summary": "Check and show changes by the current infracode",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.TerrariumCreationRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "default": "This terrarium enriches ...",
                    "example": "This terrarium enriches ..."
                },
                "name": {
                    "type": "string",
                    "default": "tr01",
                    "example": "tr01"
                }
            }
        },
        "model.TerrariumInfo": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "credentialProfile": {
                    "description": "The name of the credential profile (holder) used for this terrarium",
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "default": "This terrarium enriches ...",
                    "example": "This terrarium enriches ..."
                },
                "enrichments": {
                    "type": "string",
                    "example": "vpn/aws-to-site"
                },
                "id": {
                    "type": "string",
                    "default": "tr01",
                    "example": "tr01"
                },
                "name": {
                    "type": "string",
                    "default": "tr01",
                    "example": "tr01"
                },
                "providers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "aws",
                        "azure",
                        "gcp"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}